+++ checks/container_rt.go
@@ -108,12 +108,12 @@ func fmtContainerStats(
 
 		ifStats := ctr.Network.SumInterfaces()
 		cpus := runtime.NumCPU()
-		sys2, sys1 := ctr.CPU.SystemUsage, lastCtr.CPU.SystemUsage
+		sys2, sys1 := float64(ctr.CPU.SystemUsage), float64(lastCtr.CPU.SystemUsage)
 		chunk = append(chunk, &model.ContainerStat{
 			Id:         ctr.ID,
-			UserPct:    calculateCtrPct(uint64(ctr.CPU.User), uint64(lastCtr.CPU.User), sys2, sys1, cpus, lastRun),
-			SystemPct:  calculateCtrPct(uint64(ctr.CPU.System), uint64(lastCtr.CPU.System), sys2, sys1, cpus, lastRun),
-			TotalPct:   calculateCtrPct(uint64(ctr.CPU.User+ctr.CPU.System), uint64(lastCtr.CPU.User+lastCtr.CPU.System), sys2, sys1, cpus, lastRun),
+			UserPct:    float32(calculateCtrPct(ctr.CPU.User, lastCtr.CPU.User, sys2, sys1, cpus, lastRun)),
+			SystemPct:  float32(calculateCtrPct(ctr.CPU.System, lastCtr.CPU.System, sys2, sys1, cpus, lastRun)),
+			TotalPct:   float32(calculateCtrPct(ctr.CPU.User+ctr.CPU.System, lastCtr.CPU.User+lastCtr.CPU.System, sys2, sys1, cpus, lastRun)),
 			CpuLimit:   float32(ctr.Limits.CPULimit),
 			MemRss:     ctr.Memory.RSS,
 			MemCache:   ctr.Memory.Cache,
